Work on fixing incomplete structs.
The comment added in explains one issue fix. The other is the constant
slot reservation mechanism (not yet fixed). Constants can reserve slots
even when they're incomplete. And later uses of the declaration will
early-return because they see a slot.

We could fix this by having slots know if they are complete or not. This
feels a lot like the completeness bool we added to structs. Perhaps we
don't need both.

+++ type/primitive.h
@@ -67,7 +67,7 @@ inline Type const *Label      = &(*internal::kPrimitiveArray)[18];
 inline bool IsNumeric(Type const *t) { return t >= Nat8 and t <= Float64; }
 inline bool IsIntegral(Type const *t) { return t >= Nat8 and t <= Int64; }
 inline bool IsUnsignedNumeric(Type const *t) { return t >= Nat8 and t <= Nat64; }
-inline bool IsSignedNumeric(Type const *t) { return t >= Int8 and t <= Int64; }
+inline bool IsSignedNumeric(Type const *t) { return t >= Int8 and t <= Float64; }
 inline bool IsFloatingPoint(Type const *t) {
   return t >= Float32 and t <= Float64;
 }
